requirements.txt
autopep8==1.6.0
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.2
Flask-MySQL==1.5.2
Flask-WTF==1.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycodestyle==2.8.0
PyMySQL==1.0.2
python-decouple==3.5
python-dotenv==0.19.1
toml==0.10.2
visitor==0.1.3
Werkzeug==2.0.2
Flask-WTF==1.0.1
paho-mqtt
